"made quick changes for uploading"

+++ sahakaari/database/migrations/2020_02_03_154737_create_saving_balances_table.php
@@ -16,17 +16,15 @@ class CreateSavingBalancesTable extends Migration
         Schema::create('saving_balances', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
-            $table->foreign('user_id')
-                ->references('id')->on('users');
+            $table->foreign('user_id')->references('id')->on('users');
             $table->unsignedBigInteger('saving_id');
-            $table->foreign('saving_id')
-                ->references('id')->on('savings');
+            $table->foreign('saving_id')->references('id')->on('savings');
             $table->text('description')->charset('utf8')->collation('utf8_unicode_ci')->nullable();
-            $table->bigInteger('balance');
+            $table->decimal('balance', 12, 5);
             $table->bigInteger('withdraw')->nullable();
             $table->bigInteger('deposit')->nullable();
-            $table->bigInteger('interest_amount')->default(0);
-            $table->bigInteger('saving_amount')->default(0);
+            $table->decimal('interest_amount', 12, 5)->default(0);
+            $table->decimal('saving_amount', 12, 5)->default(0);
             $table->string('remarks')->charset('utf8')->collation('utf8_unicode_ci')->nullable();
             $table->string('creation_date')->charset('utf8')->collation('utf8_unicode_ci')->nullable();
             $table->timestamps();
